Review Hibernate 3.6.5 patch and OpenMRS upgrade patch - SEARCH-10
git-svn-id: http://svn.openmrs.org/openmrs/trunk@24959 5bac5841-c719-aa4e-b3fe-cce5062f897a

+++ api/src/main/java/org/openmrs/api/ProviderService.java
@@ -155,7 +155,7 @@ public interface ProviderService extends OpenmrsService {
 	 */
 	@Transactional(readOnly = true)
 	@Authorized( { PrivilegeConstants.VIEW_PROVIDERS })
-	public Long getCountOfProviders(String query);
+	public Integer getCountOfProviders(String query);
 	
 	/**
 	 * Gets all provider attribute types including retired provider attribute types. This method
